refactor(KlaviyoSwift): extract KlaviyoState request-building helpers into an extension (MAGE-952)
Pure relocation, no behavior change.
